requirements.txt
Django==1.6.1
django-configurations>=0.4
dj-email-url==0.0.1

ipdb>=0.8
ipython>=1.1.0
django_extensions>=1.3.2
django-debug-toolbar>=1.0.1
envdir>=0.6.1
dj-database-url==0.2.2
django-cache-url==0.7.0
south>=0.8.4
django-grappelli>=2.5.1
psycopg2>=2.5.2
pylibmc>=1.2.3
django-categories>=1.2.2
sorl-thumbnail>=11.12

